requirements.txt
behave==1.2.6
requests==2.32.3
websocket-client==1.8.0